+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Change Log
 
+## 0.2.0 (November 13, 2016)
 
+- Feature: Add linux support by xclip
+- Feature: Support use the selected text as the image name
 
 ## 0.1.0 (November 12, 2016)
 
-- Add windows support by powershell(@kivle in #2)
+- Feature: Add windows support by powershell(@kivle in #2)
 
 ## 0.0.1